+++ .github/workflows/linters.yml
@@ -7,6 +7,13 @@ AllCops:
 
 Layout/LineLength:
   Max: 120
+  Metrics/ModuleLength:
+  Enabled: false
+Metrics/CyclomaticComplexity:
+  Enabled: false
+Metrics/PerceivedComplexity:
+  Enabled: false
+
 Metrics/MethodLength:
   Max: 20
 Metrics/AbcSize:
@@ -47,3 +54,6 @@ Style/HashTransformKeys:
   Enabled: false
 Style/HashTransformValues:
   Enabled: false
+  Style/CaseEquality:
+  Enabled: false
+
